requirements.txt
asgiref==3.9.1
astroid==3.3.11
colorama==0.4.6
dill==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.1
tzdata==2025.2
